refactor: replace TechClock component with improved structure and styling
