+++ src/components/CurrentTime.tsx
@@ -1,0 +1,32 @@
+import { useEffect, useState } from "react";
+import { Clock } from "lucide-react";
+
+export default function TechClock() {
+  const [time, setTime] = useState(new Date());
+
+  useEffect(() => {
+    const timer = setInterval(() => {
+      setTime(new Date());
+    }, 1000);
+
+    return () => clearInterval(timer);
+  }, []);
+
+  const formatDateTime = (date: Date) => {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, "0");
+    const day = String(date.getDate()).padStart(2, "0");
+    const hours = String(date.getHours()).padStart(2, "0");
+    const minutes = String(date.getMinutes()).padStart(2, "0");
+    const seconds = String(date.getSeconds()).padStart(2, "0");
+
+    return `${year}/${month}/${day} ${hours}:${minutes}:${seconds}`;
+  };
+
+  return (
+    <div className="flex gap-2 items-center font-mono text-sm text-white/70">
+      <Clock className="size-4" />
+      <span>{formatDateTime(time)}</span>
+    </div>
+  );
+}
