--- src/components/TechClock.css
@@ -1,36 +1,0 @@
-.tech-clock {
-  position: relative;
-  font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-}
-
-.time-display {
-  display: flex;
-  align-items: center;
-  background: rgba(30, 50, 90, 0.9);
-  backdrop-filter: blur(8px);
-  border: 1px solid rgba(59, 130, 246, 0.3);
-  border-radius: 0.5rem;
-  padding: 0.5rem 0.75rem;
-  box-shadow: 0 2px 8px rgba(59, 130, 246, 0.1);
-  transition: all 0.2s ease;
-}
-
-.time-display:hover {
-  border-color: rgba(59, 130, 246, 0.5);
-  box-shadow: 0 4px 12px rgba(59, 130, 246, 0.15);
-}
-
-.time-icon {
-  width: 0.9rem;
-  height: 0.9rem;
-  color: #60a5fa;
-  margin-right: 0.5rem;
-}
-
-.time-text {
-  color: #dbeafe;
-  font-size: 0.9rem;
-  font-weight: 500;
-  letter-spacing: 0.025em;
-}
-
