--- src/components/TechClock.tsx
@@ -1,35 +1,0 @@
-import { useEffect, useState } from 'react';
-import { Clock } from 'lucide-react';
-import './TechClock.css';
-
-export default function TechClock() {
-  const [time, setTime] = useState(new Date());
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setTime(new Date());
-    }, 1000);
-
-    return () => clearInterval(timer);
-  }, []);
-
-  const formatDateTime = (date: Date) => {
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    const hours = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-    const seconds = String(date.getSeconds()).padStart(2, '0');
-
-    return `${year}/${month}/${day} ${hours}:${minutes}:${seconds}`;
-  };
-
-  return (
-    <div className="tech-clock">
-      <div className="time-display">
-        <Clock className="time-icon" />
-        <span className="time-text">{formatDateTime(time)}</span>
-      </div>
-    </div>
-  );
-}
